Feat article edit - url,html, view 생성

+++ articles/views.py
@@ -48,5 +48,17 @@ def detail(request, pk):
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
+    
+    if request.method == 'POST':
+        article.delete()
+        
     return redirect('articles:dinner')
+
+def edit(requset, pk):
+    article = Article.objects.get(pk=pk)
+    
+    context = {
+        'article': article,
+    }
+    
+    return render(request, 'edit.html', context)
